chore: move investor-yearn into lib
