+++ package.json
@@ -102,7 +102,6 @@
     "@shapeshiftoss/hdwallet-xdefi": "^1.48.0",
     "@shapeshiftoss/investor-foxy": "workspace:^",
     "@shapeshiftoss/investor-idle": "workspace:^",
-    "@shapeshiftoss/investor-yearn": "workspace:^",
     "@shapeshiftoss/logger": "workspace:^",
     "@shapeshiftoss/types": "workspace:^",
     "@shapeshiftoss/unchained-client": "workspace:^",
