--- packages/investor-yearn/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../tsconfig.packages.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "include": ["src/**/*"],
-  "exclude": ["dist"],
-  "references": [
-    { "path": "../caip" },
-    { "path": "../chain-adapters" },
-    { "path": "../investor" },
-    { "path": "../logger" },
-    { "path": "../types" },
-  ]
-}
