+++ src/features/defi/contexts/YearnProvider/YearnProvider.tsx
@@ -1,12 +1,13 @@
-import type { YearnInvestor } from '@shapeshiftoss/investor-yearn'
 import { KnownChainIds } from '@shapeshiftoss/types'
 import { getYearnInvestor } from 'features/defi/contexts/YearnProvider/yearnInvestorSingleton'
 import type { PropsWithChildren } from 'react'
 import React, { useContext, useEffect, useState } from 'react'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
+import type { YearnInvestor } from 'lib/investor/investor-yearn'
 import { logger } from 'lib/logger'
 import { selectFeatureFlags } from 'state/slices/selectors'
 import { store } from 'state/store'
+
 const moduleLogger = logger.child({ namespace: ['YearnProvider'] })
 
 type YearnContextProps = {
