+++ src/lib/investor/investor-yearn/YearnInvestor.ts
