+++ src/lib/investor/investor-yearn/YearnOpportunity.ts
