+++ tsconfig.json
@@ -42,7 +42,6 @@
     { "path": "./packages/investor" },
     { "path": "./packages/investor-foxy" },
     { "path": "./packages/investor-idle" },
-    { "path": "./packages/investor-yearn" },
     { "path": "./packages/logger" },
     { "path": "./packages/types" },
     { "path": "./packages/unchained-client" },
